fix: fixed errors in formatting and added functions
