+++ js/ads.js
@@ -1,4 +1,4 @@
-// Реализация объявлений и их сравнение
+// Реализация объявлений
 
 // Создание массива из десяти сгенерированных объектов
